+++ src/views/account-details/components/transaction-amount/transaction-amount.view.tsx
