+++ src/views/account-details/components/transaction-type/transaction-type.view.tsx
@@ -1,16 +1,24 @@
 import React from "react";
-import PropTypes from "prop-types";
 import { TxType } from "@hermeznetwork/hermezjs/src/enums";
 
-import depositedIcon from "../../../../images/icons/deposited.svg";
-import receivedIcon from "../../../../images/icons/received.svg";
-import sentIcon from "../../../../images/icons/sent.svg";
-import withdrawnIcon from "../../../../images/icons/withdrawn.svg";
+import depositedIcon from "src/images/icons/deposited.svg";
+import receivedIcon from "src/images/icons/received.svg";
+import sentIcon from "src/images/icons/sent.svg";
+import withdrawnIcon from "src/images/icons/withdrawn.svg";
 
-function TransactionType({ type, fromAccountIndex, accountIndex }) {
+interface TransactionTypeProps {
+  type: string;
+  fromAccountIndex: string;
+  accountIndex: string;
+}
+
+function TransactionType({
+  type,
+  fromAccountIndex,
+  accountIndex,
+}: TransactionTypeProps): JSX.Element {
   /**
    * Returns the icon corresponding to the transaction type
-   * @returns {string} - Icon content encoded in base64
    */
   function getIcon() {
     switch (type) {
@@ -38,11 +46,7 @@ function TransactionType({ type, fromAccountIndex, accountIndex }) {
     }
   }
 
-  return <img src={getIcon(type)} alt="Transaction type" />;
+  return <img src={getIcon()} alt="Transaction type" />;
 }
-
-TxType.propTypes = {
-  type: PropTypes.string.isRequired,
-};
 
 export default TransactionType;
