+++ src/views/home/components/pending-deposit-list/pending-deposit-list.styles.ts
@@ -1,6 +1,8 @@
 import { createUseStyles } from "react-jss";
 
-const usePendingDepositListStyles = createUseStyles((theme) => ({
+import { Theme } from "src/styles/theme";
+
+const usePendingDepositListStyles = createUseStyles((theme: Theme) => ({
   pendingDeposit: {
     width: "100%",
     marginBottom: theme.spacing(2.5),
