--- src/views/home/components/pending-deposit-list/pending-deposit-list.view.jsx
@@ -1,46 +1,0 @@
-import React from "react";
-
-import {
-  getFixedTokenAmount,
-  getTokenAmountInPreferredCurrency,
-} from "../../../../utils/currencies";
-import Account from "../../../shared/account/account.view";
-import usePendingDepositListStyles from "./pending-deposit-list.styles";
-
-function PendingDepositList({
-  deposits,
-  preferredCurrency,
-  fiatExchangeRates,
-  coordinatorState,
-  onAccountClick,
-}) {
-  const classes = usePendingDepositListStyles();
-
-  return (
-    <>
-      {deposits.map((deposit) => {
-        return (
-          <div key={deposit.hash} className={classes.pendingDeposit}>
-            <Account
-              hasPendingDeposit
-              balance={deposit.amount}
-              fiatBalance={getTokenAmountInPreferredCurrency(
-                getFixedTokenAmount(deposit.amount, deposit.token.decimals),
-                deposit.token.USD,
-                preferredCurrency,
-                fiatExchangeRates
-              )}
-              token={deposit.token}
-              preferredCurrency={preferredCurrency}
-              coordinatorState={coordinatorState}
-              timestamp={deposit.timestamp}
-              onClick={onAccountClick}
-            />
-          </div>
-        );
-      })}
-    </>
-  );
-}
-
-export default PendingDepositList;
